Amenity Auto-Repair (#594)
* working proof of concept for ntu request/flow

* ntu is now measured in Float rather than Int; adjusted the type declarations in ChatActor, as it was complaining

* added auto repair information for amenities; modified terminal working example to include periodic repair timer

* crude power state acknowledgement exchange between ntu silo and building, following by building and amenities; better management and control over auto-repair

* separated auto-repair from terminal control; resource silo publishes an appropriate power state message upon formal start-up

* various Damageble objects have been granted auto-repair; doors no longer report no event for intercepting unhandled messages

* documentation and comments; unit tests and integration tests for AmneityAutoRepair, ResourceSilo, and BuildingActor

* merge plus some fixes

* addition checks for auto-repairing cavern equipment and for auto-repairing un-owned equipment

+++ src/main/scala/net/psforever/objects/serverobject/resourcesilo/ResourceSilo.scala
@@ -16,7 +16,7 @@ class ResourceSilo extends Amenity with CommonNtuContainer {
   // For the NTU display bar
   private var capacitorDisplay: Long = 0
 
-  def MaxNtuCapacitor : Int = Definition.MaxNtuCapacitor
+  def MaxNtuCapacitor : Float = Definition.MaxNtuCapacitor
 
   def LowNtuWarningOn: Boolean = lowNtuWarningOn
   def LowNtuWarningOn_=(enabled: Boolean): Boolean = {
@@ -24,7 +24,7 @@ class ResourceSilo extends Amenity with CommonNtuContainer {
     LowNtuWarningOn
   }
 
-  def CapacitorDisplay : Long = scala.math.ceil((NtuCapacitor.toFloat / MaxNtuCapacitor.toFloat) * 10).toInt
+  def CapacitorDisplay : Long = scala.math.ceil((NtuCapacitor / MaxNtuCapacitor) * 10).toInt
 
   def Definition: ResourceSiloDefinition = GlobalDefinitions.resource_silo
 
@@ -34,7 +34,7 @@ class ResourceSilo extends Amenity with CommonNtuContainer {
 }
 
 object ResourceSilo {
-  final case class UpdateChargeLevel(amount: Int)
+  final case class UpdateChargeLevel(amount: Float)
   final case class LowNtuWarning(enabled: Boolean)
   sealed trait Exchange
   final case class ChargeEvent() extends Exchange
